rename file-manager folder, didn't like the hyphen
