+++ src/features/parser/combinedContainer.js
@@ -7,7 +7,7 @@ import CompiledOutput from './compiledOutput'
 import MessagePanel from './messagePanel'
 import MobilePage from '../common/mobilePage'
 import ControlsContainer from '../parser/controlsContainer'
-import FileManagerContainer from '../file-manager/fileManagerContainer'
+import FileManagerContainer from '../fileManager/fileManagerContainer'
 
 import { space } from '../common/theme'
